+++ utils/magic/magic.go
@@ -1,0 +1,66 @@
+package magic
+
+import (
+	"fmt"
+	"strings"
+	"time"
+)
+
+// 渲染魔法变量，获取渲染结果
+func GoMagic(rawSql string, now time.Time) (ret string) {
+	sps := strings.Split(rawSql, "@(") //@()，对于每个分片找右括号
+	ret = sps[0]
+	for idx := 1; idx < len(sps); idx++ {
+		idxr := strings.Index(sps[idx], ")")
+		if idxr == -1 {
+			return rawSql
+		}
+		ret += convertMagic(sps[idx][0:idxr], now)
+		if idxr+1 < len(sps[idx]) {
+			ret += sps[idx][idxr+1:]
+		}
+	}
+
+	return ret
+}
+
+// 支持 YYYY, YY, MM, DD, hh, mm, ss, M, D, h, m, s
+func convertMagic(magic string, now time.Time) (ret string) {
+	switch magic {
+	case "YYYY":
+		return fmt.Sprintf("%d", now.Year())
+	case "YY":
+		return fmt.Sprintf("%d", now.Year())[2:]
+	case "MM":
+		m := int(now.Month())
+		return fmt.Sprintf("%02d", m)
+	case "M":
+		m := int(now.Month())
+		return fmt.Sprintf("%d", m)
+	case "D":
+		d := int(now.Day())
+		return fmt.Sprintf("%d", d)
+	case "DD":
+		d := int(now.Day())
+		return fmt.Sprintf("%02d", d)
+	case "hh":
+		h := now.Hour()
+		return fmt.Sprintf("%02d", h)
+	case "h":
+		h := now.Hour()
+		return fmt.Sprintf("%d", h)
+	case "mm":
+		m := now.Minute()
+		return fmt.Sprintf("%02d", m)
+	case "m":
+		m := now.Minute()
+		return fmt.Sprintf("%d", m)
+	case "ss":
+		s := now.Second()
+		return fmt.Sprintf("%02d", s)
+	case "s":
+		s := now.Second()
+		return fmt.Sprintf("%d", s)
+	}
+	return ""
+}
